Code clean up done
