+++ app/src/main/java/digital/lamp/mindlamp/utils/AppConstants.kt
@@ -12,7 +12,7 @@ internal object AppConstants {
     const val NOTIFICATION_SURVEY_OPEN = "lamp_survey_open"
     const val NOTIFICATION_ID = 101
     const val INITIAL_TRIGGER = 120000L //30 seconds
-    const val ALARM_INTERVAL = 300000L //10 Minutes
+    const val ALARM_INTERVAL = 60000L //10 Minutes
 
     const val REQUEST_ID_MULTIPLE_PERMISSIONS = 1
