+++ sensor-core/src/main/AndroidManifest.xml
@@ -89,110 +89,6 @@
             android:name="com.mindlamp.SignificantMotion"
             android:exported="true" />
 
-        <!-- Core providers -->
-        <provider
-            android:name="com.mindlamp.providers.Scheduler_Provider"
-            android:authorities="${applicationId}.provider.scheduler"
-            android:exported="true"
-            android:readPermission="com.aware.READ_CONTEXT_DATA"
-            android:writePermission="com.aware.WRITE_CONTEXT_DATA" />
-        <provider
-            android:name="com.mindlamp.providers.Locations_Provider"
-            android:authorities="${applicationId}.provider.locations"
-            android:exported="true"
-            android:label="@string/sensor_locations"
-            android:readPermission="com.aware.READ_CONTEXT_DATA"
-            android:syncable="true"
-            android:writePermission="com.aware.WRITE_CONTEXT_DATA" />
-        <provider
-            android:name="com.mindlamp.providers.Linear_Accelerometer_Provider"
-            android:authorities="${applicationId}.provider.accelerometer.linear"
-            android:exported="true"
-            android:label="@string/sensor_linear_accelerometer"
-            android:readPermission="com.aware.READ_CONTEXT_DATA"
-            android:syncable="true"
-            android:writePermission="com.aware.WRITE_CONTEXT_DATA" />
-        <provider
-            android:name="com.mindlamp.providers.Barometer_Provider"
-            android:authorities="${applicationId}.provider.barometer"
-            android:exported="true"
-            android:label="Barometer"
-            android:readPermission="com.aware.READ_CONTEXT_DATA"
-            android:syncable="true"
-            android:writePermission="com.aware.WRITE_CONTEXT_DATA" />
-        <provider
-            android:name="com.mindlamp.providers.Magnetometer_Provider"
-            android:authorities="${applicationId}.provider.magnetometer"
-            android:exported="true"
-            android:label="@string/sensor_magnetometer"
-            android:readPermission="com.aware.READ_CONTEXT_DATA"
-            android:syncable="true"
-            android:writePermission="com.aware.WRITE_CONTEXT_DATA" />
-        <provider
-            android:name="com.mindlamp.providers.Light_Provider"
-            android:authorities="${applicationId}.provider.light"
-            android:exported="true"
-            android:label="@string/sensor_light"
-            android:readPermission="com.aware.READ_CONTEXT_DATA"
-            android:syncable="true"
-            android:writePermission="com.aware.WRITE_CONTEXT_DATA" />
-        <provider
-            android:name="com.mindlamp.providers.Lamp_Provider"
-            android:authorities="${applicationId}.provider.aware"
-            android:exported="true"
-            android:label="@string/sensor_aware"
-            android:readPermission="com.aware.READ_CONTEXT_DATA"
-            android:syncable="true"
-            android:writePermission="com.aware.WRITE_CONTEXT_DATA" />
-        <provider
-            android:name="com.mindlamp.providers.Rotation_Provider"
-            android:authorities="${applicationId}.provider.rotation"
-            android:exported="true"
-            android:label="@string/sensor_rotation"
-            android:readPermission="com.aware.READ_CONTEXT_DATA"
-            android:syncable="true"
-            android:writePermission="com.aware.WRITE_CONTEXT_DATA" />
-        <provider
-            android:name="com.mindlamp.providers.WiFi_Provider"
-            android:authorities="${applicationId}.provider.wifi"
-            android:exported="true"
-            android:label="@string/sensor_wifi"
-            android:readPermission="com.aware.READ_CONTEXT_DATA"
-            android:syncable="true"
-            android:writePermission="com.aware.WRITE_CONTEXT_DATA" />
-        <provider
-            android:name="com.mindlamp.providers.Accelerometer_Provider"
-            android:authorities="${applicationId}.provider.accelerometer"
-            android:exported="true"
-            android:label="@string/sensor_accelerometer"
-            android:readPermission="com.aware.READ_CONTEXT_DATA"
-            android:syncable="true"
-            android:writePermission="com.aware.WRITE_CONTEXT_DATA" />
-        <provider
-            android:name="com.mindlamp.providers.Gyroscope_Provider"
-            android:authorities="${applicationId}.provider.gyroscope"
-            android:exported="true"
-            android:label="@string/sensor_gyroscope"
-            android:readPermission="com.aware.READ_CONTEXT_DATA"
-            android:syncable="true"
-            android:writePermission="com.aware.WRITE_CONTEXT_DATA" />
-        <provider
-            android:name="com.mindlamp.providers.Screen_Provider"
-            android:authorities="${applicationId}.provider.screen"
-            android:exported="true"
-            android:label="@string/sensor_screen"
-            android:readPermission="com.aware.READ_CONTEXT_DATA"
-            android:syncable="true"
-            android:writePermission="com.aware.WRITE_CONTEXT_DATA" />
-        <provider
-            android:name="com.mindlamp.providers.Significant_Provider"
-            android:authorities="${applicationId}.provider.significant"
-            android:exported="true"
-            android:label="@string/sensor_significant"
-            android:readPermission="com.aware.READ_CONTEXT_DATA"
-            android:syncable="true"
-            android:writePermission="com.aware.WRITE_CONTEXT_DATA" />
-
         <provider
             android:name="androidx.core.content.FileProvider"
             android:authorities="${applicationId}.provider.storage"
