+++ sensor-core/src/main/java/com/mindlamp/Barometer.java
@@ -2,26 +2,21 @@
 package com.mindlamp;
 
 import android.content.BroadcastReceiver;
-import android.content.ContentResolver;
 import android.content.ContentValues;
 import android.content.Context;
 import android.content.Intent;
 import android.content.IntentFilter;
-import android.content.SyncRequest;
 import android.hardware.Sensor;
 import android.hardware.SensorEvent;
 import android.hardware.SensorEventListener;
 import android.hardware.SensorManager;
-import android.os.Bundle;
 import android.os.Handler;
 import android.os.HandlerThread;
 import android.os.IBinder;
 import android.os.PowerManager;
+import android.provider.BaseColumns;
 import android.util.Log;
 
-import com.mindlamp.providers.Barometer_Provider;
-import com.mindlamp.providers.Barometer_Provider.Barometer_Data;
-import com.mindlamp.providers.Barometer_Provider.Barometer_Sensor;
 import com.mindlamp.utils.LampConstants;
 import com.mindlamp.utils.Lamp_Sensor;
 
@@ -54,7 +49,6 @@ public class Barometer extends Lamp_Sensor implements SensorEventListener {
     private static int FREQUENCY = -1;
     private static double THRESHOLD = 0;
     // Reject any data points that come in more often than frequency
-    private static boolean ENFORCE_FREQUENCY = false;
 
     public static final String ACTION_LAMP_BAROMETER = "ACTION_LAMP_BAROMETER";
     public static final String ACTION_LAMP_BAROMETER_LABEL = "ACTION_LAMP_BAROMETER_LABEL";
@@ -146,8 +140,6 @@ public class Barometer extends Lamp_Sensor implements SensorEventListener {
     public void onCreate() {
         super.onCreate();
 
-        AUTHORITY = Barometer_Provider.getAuthority(this);
-
         mSensorManager = (SensorManager) getSystemService(SENSOR_SERVICE);
 
         mPressure = mSensorManager.getDefaultSensor(Sensor.TYPE_PRESSURE);
@@ -219,5 +211,15 @@ public class Barometer extends Lamp_Sensor implements SensorEventListener {
     @Override
     public IBinder onBind(Intent intent) {
         return null;
+    }
+
+    public static final class Barometer_Data implements BaseColumns {
+
+        public static final String _ID = "_id";
+        public static final String TIMESTAMP = "timestamp";
+        public static final String DEVICE_ID = "device_id";
+        public static final String AMBIENT_PRESSURE = "double_values_0";
+        public static final String ACCURACY = "accuracy";
+        public static final String LABEL = "label";
     }
 }
