+++ sensor-core/src/main/java/com/mindlamp/Gyroscope.java
@@ -14,8 +14,8 @@ import android.os.Handler;
 import android.os.HandlerThread;
 import android.os.IBinder;
 import android.os.PowerManager;
+import android.provider.BaseColumns;
 import android.util.Log;
-import com.mindlamp.providers.Gyroscope_Provider.Gyroscope_Data;
 import com.mindlamp.utils.LampConstants;
 import com.mindlamp.utils.Lamp_Sensor;
 
@@ -249,5 +249,17 @@ public class Gyroscope extends Lamp_Sensor implements SensorEventListener {
     @Override
     public IBinder onBind(Intent intent) {
         return null;
+    }
+
+    public static final class Gyroscope_Data implements BaseColumns {
+
+        public static final String _ID = "_id";
+        public static final String TIMESTAMP = "timestamp";
+        public static final String DEVICE_ID = "device_id";
+        public static final String VALUES_0 = "double_values_0";
+        public static final String VALUES_1 = "double_values_1";
+        public static final String VALUES_2 = "double_values_2";
+        public static final String ACCURACY = "accuracy";
+        public static final String LABEL = "label";
     }
 }
