+++ sensor-core/src/main/java/com/mindlamp/LinearAccelerometer.java
@@ -2,24 +2,20 @@
 package com.mindlamp;
 
 import android.content.BroadcastReceiver;
-import android.content.ContentResolver;
 import android.content.ContentValues;
 import android.content.Context;
 import android.content.Intent;
 import android.content.IntentFilter;
-import android.content.SyncRequest;
 import android.hardware.Sensor;
 import android.hardware.SensorEvent;
 import android.hardware.SensorEventListener;
 import android.hardware.SensorManager;
-import android.os.Bundle;
 import android.os.Handler;
 import android.os.HandlerThread;
 import android.os.IBinder;
 import android.os.PowerManager;
+import android.provider.BaseColumns;
 import android.util.Log;
-import com.mindlamp.providers.Linear_Accelerometer_Provider;
-import com.mindlamp.providers.Linear_Accelerometer_Provider.Linear_Accelerometer_Data;
 import com.mindlamp.utils.LampConstants;
 import com.mindlamp.utils.Lamp_Sensor;
 
@@ -183,7 +179,7 @@ public class LinearAccelerometer extends Lamp_Sensor implements SensorEventListe
     public void onCreate() {
         super.onCreate();
 
-        AUTHORITY = Linear_Accelerometer_Provider.getAuthority(this);
+        AUTHORITY = getPackageName() + ".provider.accelerometer.linear";
 
         mSensorManager = (SensorManager) getSystemService(SENSOR_SERVICE);
         mLinearAccelerator = mSensorManager.getDefaultSensor(Sensor.TYPE_LINEAR_ACCELERATION);
@@ -256,5 +252,17 @@ public class LinearAccelerometer extends Lamp_Sensor implements SensorEventListe
     @Override
     public IBinder onBind(Intent intent) {
         return null;
+    }
+
+    public static final class Linear_Accelerometer_Data implements BaseColumns {
+
+        public static final String _ID = "_id";
+        public static final String TIMESTAMP = "timestamp";
+        public static final String DEVICE_ID = "device_id";
+        public static final String VALUES_0 = "double_values_0";
+        public static final String VALUES_1 = "double_values_1";
+        public static final String VALUES_2 = "double_values_2";
+        public static final String ACCURACY = "accuracy";
+        public static final String LABEL = "label";
     }
 }
