+++ sensor-core/src/main/java/com/mindlamp/Magnetometer.java
@@ -14,8 +14,8 @@ import android.os.Handler;
 import android.os.HandlerThread;
 import android.os.IBinder;
 import android.os.PowerManager;
+import android.provider.BaseColumns;
 import android.util.Log;
-import com.mindlamp.providers.Magnetometer_Provider.Magnetometer_Data;
 import com.mindlamp.utils.LampConstants;
 import com.mindlamp.utils.Lamp_Sensor;
 
@@ -50,7 +50,6 @@ public class Magnetometer extends Lamp_Sensor implements SensorEventListener {
     private static int FREQUENCY = -1;
     private static double THRESHOLD = 0;
     // Reject any data points that come in more often than frequency
-    private static boolean ENFORCE_FREQUENCY = false;
 
     /**
      * Broadcasted event: new sensor values
@@ -236,5 +235,17 @@ public class Magnetometer extends Lamp_Sensor implements SensorEventListener {
     @Override
     public IBinder onBind(Intent intent) {
         return null;
+    }
+
+    public static final class Magnetometer_Data implements BaseColumns {
+
+        public static final String _ID = "_id";
+        public static final String TIMESTAMP = "timestamp";
+        public static final String DEVICE_ID = "device_id";
+        public static final String VALUES_0 = "double_values_0";
+        public static final String VALUES_1 = "double_values_1";
+        public static final String VALUES_2 = "double_values_2";
+        public static final String ACCURACY = "accuracy";
+        public static final String LABEL = "label";
     }
 }
