+++ sensor-core/src/main/java/com/mindlamp/SignificantMotion.java
@@ -10,6 +10,7 @@ import android.os.Handler;
 import android.os.HandlerThread;
 import android.os.IBinder;
 import android.os.PowerManager;
+import android.provider.BaseColumns;
 import android.util.Log;
 
 import com.mindlamp.providers.Significant_Provider;
